Update Cartthrob_shipping_ups.php
Modified to include setting to allow carts with ZERO weight to have ZERO shipping applied to them in the case that checkout may include 100% weightless (digital) items in some cases, but not in others. By default this plugin assumes a weight of 1 lb or more. This setting will make it possible to assume that some checkouts will not need a shipping charge.